requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
numpy==1.24.4
cython==0.29.36
blis==0.7.9
murmurhash==1.0.9
cymem==2.0.6
preshed==3.0.6
spacy==3.6.1
pymupdf==1.23.7
openai==0.28
pymysql==1.1.0
Werkzeug==3.0.3
sentence-transformers
scikit-learn